+++ src/lib/common/common.cu
@@ -41,8 +41,8 @@ namespace gpuntt
         return chk;
     }
 
-    template bool check_result<unsigned long long>(unsigned long long* input1,
-                                                   unsigned long long* input2,
+    template bool check_result<unsigned long>(unsigned long* input1,
+                                                   unsigned long* input2,
                                                    int size);
 
     template bool check_result<unsigned>(unsigned* input1, unsigned* input2,
